Added option to display RX SNR dB for CRSF instead of RSSI dBm.

+++ src/main/osd/osd.h
@@ -278,7 +278,7 @@ typedef struct osdConfig_s {
     uint8_t overlay_radio_mode;
     char profile[OSD_PROFILE_COUNT][OSD_PROFILE_NAME_LENGTH + 1];
     uint16_t link_quality_alarm;
-    uint8_t rssi_dbm_alarm;
+    int16_t rssi_dbm_alarm;
     uint8_t gps_sats_show_hdop;
     int8_t rcChannels[OSD_RCCHANNELS_COUNT];  // RC channel values to display, -1 if none
     uint8_t displayPortDevice;                // osdDisplayPortDevice_e
@@ -310,7 +310,7 @@ typedef struct statistic_s {
     int16_t max_esc_temp;
     int32_t max_esc_rpm;
     uint16_t min_link_quality;
-    uint8_t min_rssi_dbm;
+    int16_t min_rssi_dbm;
 } statistic_t;
 
 extern timeUs_t resumeRefreshAt;
